+++ framework/conf/extensions/maint.defaults.php
@@ -24,6 +24,7 @@ if (!defined('MAINTENANCE_RETURN_TIME')) define('MAINTENANCE_RETURN_TIME','0');
 
 if (!defined('DEVELOPMENT')) define('DEVELOPMENT','0');
 if (!defined('SMARTY_DEVELOPMENT')) define('SMARTY_DEVELOPMENT','0');
+if (!defined('XMLRPC_DEVELOPMENT')) define('XMLRPC_DEVELOPMENT','0');
 if (!defined('SMARTY_CACHING')) define('SMARTY_CACHING','0');
 if (!defined('LOGGER')) define('LOGGER','0');
 if (!defined('DEBUG_HISTORY')) define('DEBUG_HISTORY','0');