+++ framework/core/subsystems/expMail.php
@@ -68,9 +68,13 @@ class expMail {
 	 * @todo add further documentation for using settings other than the system default
 	 */
 	function __construct($params = array()) {
-        require_once(SWIFT_LEXER_PATH . 'AbstractLexer.php');
-        require_once(SWIFT_EMAIL_PATH . 'EmailValidator.php');
-        Egulias\EmailValidator\EmailValidator::registerAutoload();
+        if (version_compare(SWIFT_VERSION, '6.0.0', 'ge')) {
+            require_once(SWIFT_LEXER_PATH . 'AbstractLexer.php');
+            require_once(SWIFT_DEPRECATIONS_PATH . 'Deprecation.php');
+            require_once(SWIFT_LEXER_PATH . 'Token.php');
+            require_once(SWIFT_EMAIL_PATH . 'EmailValidator.php');
+            Egulias\EmailValidator\EmailValidator::registerAutoload();  // our custom autoload function
+        }
 		require_once(SWIFT_PATH . 'swift_required.php');
 
 		if (array_key_exists('method', $params)) {
@@ -95,14 +99,21 @@ class expMail {
 					}
 					break;
 				case "native":
-
 					if (isset($params['connections']) && !is_array($params['connections']) && $params['connections'] != '') {
 						// Allow custom mail parameters.
 //						$this->transport = Swift_MailTransport::newInstance($params['connections']);
-                        $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                        if (version_compare(SWIFT_VERSION, '6.0.0', 'ge')) {
+                            $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                        } else {
+                            $this->transport = new Swift_MailTransport($params['connections']);  //fixme DEPRECATED
+                        }
 					} else {
 //						$this->transport = Swift_MailTransport::newInstance();
-                        $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                        if (version_compare(SWIFT_VERSION, '6.0.0', 'ge')) {
+                            $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                        } else {
+                            $this->transport = new Swift_MailTransport($params['connections']);  //fixme DEPRECATED
+                        }
 					}
 					break;
 				case "exim":
@@ -127,11 +138,19 @@ class expMail {
 			if (isset($params['connections']) && !is_array($params['connections']) && $params['connections'] != '') {
 				// Allow custom mail parameters.
 //				$this->transport = Swift_MailTransport::newInstance($params['connections']);
-                $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                if (version_compare(SWIFT_VERSION, '6.0.0', 'ge')) {
+                    $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                } else {
+                    $this->transport = new Swift_MailTransport($params['connections']);  //fixme DEPRECATED
+                }
 			} else {
 				// Use default Mail parameters.
 //				$this->transport = Swift_MailTransport::newInstance();
-                $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                if (version_compare(SWIFT_VERSION, '6.0.0', 'ge')) {
+                    $this->transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
+                } else {
+                    $this->transport = new Swift_MailTransport($params['connections']);  //fixme DEPRECATED
+                }
 			}
 		} else {
 			/*