+++ xmlrpc.php
@@ -98,7 +98,8 @@ function exp_getModuleInstancesByType($type = null)
             $modules[$ref->source][] = $mod;
         }
     }
-//    eLog($modules, 'Modules');
+    if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1)
+        eLog($modules, 'Modules');
     return $modules;
 }
 
@@ -144,7 +145,8 @@ function getUsersBlogs($xmlrpcmsg)
                 );
             }
         }
-//eLog($structArray, 'getUsersBlogs');
+        if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1)
+            eLog($structArray, 'getUsersBlogs');
         return new xmlrpcresp(new xmlrpcval($structArray, 'array'));
     } else {
         return new xmlrpcresp(0, 1, "Login Failed");
@@ -239,7 +241,8 @@ if ($src == "1") $src = "@random48e7cdc1a672d";
 
             $post->update($params);
 
-//eLog($post->id, 'newPost');
+            if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1)
+                eLog($post->id, 'newPost');
             return new xmlrpcresp(
                 new xmlrpcval($post->id, 'string')
             ); // Return the id of the post just inserted into the DB. See mysql_insert_id() in the PHP manual.
@@ -341,7 +344,8 @@ function editPost($xmlrpcmsg)
 
             $post->update($params);
 
-//eLog($postid, 'editPost');
+            if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1)
+                eLog($postid, 'editPost');
             return new xmlrpcresp(new xmlrpcval(true, 'boolean'));
         } else {
             return new xmlrpcresp(0, 1, "Login Failed");
@@ -389,9 +393,11 @@ function getPost($xmlrpcmsg)
             foreach ($post->expTag as $tag) {
                 $selectedtags .= $tag->title . ', ';
             }
-//eLog($postid, 'getPost id');
-//$tmp = new xmlrpcval(expString::convertSmartQuotes($post->body), 'string');
-//eLog($tmp, 'getPost');
+            if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1) {
+                eLog($postid, 'getPost id');
+                $tmp = new xmlrpcval(expString::convertSmartQuotes($post->body), 'string');
+                eLog($tmp, 'getPost');
+            }
             return new xmlrpcresp(
                 new xmlrpcval(
                     array(
@@ -452,7 +458,8 @@ function deletePost($xmlrpcmsg)
             $post->delete();
         }
 
-//eLog($postid, 'deletePost');
+        if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1)
+            eLog($postid, 'deletePost');
         return new xmlrpcresp(new xmlrpcval(true, 'boolean'));
     } else {
         return new xmlrpcresp(0, 1, "Login Failed");
@@ -535,8 +542,10 @@ if ($src == "1") $src = "@random48e7cdc1a672d";
                     );
                 }
             }
-//eLog($src, 'getRecentPosts src');
-//eLog($structArray, 'getRecentPosts');
+            if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1) {
+                eLog($src, 'getRecentPosts src');
+                eLog($structArray, 'getRecentPosts');
+            }
             return new xmlrpcresp(new xmlrpcval($structArray, 'array')); // Return type is struct[] (array of struct)
         } else {
             return new xmlrpcresp(0, 1, "Login Failed");
@@ -582,8 +591,10 @@ if ($src == "1") $src = "@random48e7cdc1a672d";
                 );
             }
         }
-//eLog($src, 'getCategories src');
-//eLog($structArray, 'getCategories');
+        if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1) {
+            eLog($src, 'getCategories src');
+            eLog($structArray, 'getCategories');
+        }
         return new xmlrpcresp(new xmlrpcval($structArray, 'array')); // Return type is struct[] (array of struct)
     } else {
         return new xmlrpcresp(0, 1, 'Login Failed');
@@ -613,7 +624,8 @@ if ($src == "1") $src = "@random48e7cdc1a672d";
                 );
             }
         }
-//eLog($structArray, 'getCategories_mt');
+        if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1)
+            eLog($structArray, 'getCategories_mt');
         return new xmlrpcresp(new xmlrpcval($structArray, 'array')); // Return type is struct[] (array of struct)
     } else {
         return new xmlrpcresp(0, 1, 'Login Failed');
@@ -659,8 +671,10 @@ if ($src == "1") $src = "@random48e7cdc1a672d";
                 }
             }
         }
-//eLog($src, 'getTerms src');
-//eLog($structArray, 'getTerms');
+        if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1) {
+            eLog($src, 'getTerms src');
+            eLog($structArray, 'getTerms');
+        }
         return new xmlrpcresp(new xmlrpcval($structArray, 'array')); // Return type is struct[] (array of struct)
     } else {
         return new xmlrpcresp(0, 1, 'Login Failed');
@@ -699,7 +713,8 @@ if ($src == "1") $src = "@random48e7cdc1a672d";
                 ), 'struct'
             );
         }
-//eLog($structArray, 'getAuthors');
+        if (defined('XMLRPC_DEVELOPMENT') && XMLRPC_DEVELOPMENT == 1)
+            eLog($structArray, 'getAuthors');
         return new xmlrpcresp(new xmlrpcval($structArray, 'array')); // Return type is struct[] (array of struct)
     } else {
         return new xmlrpcresp(0, 1, 'Login Failed');